% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_str.R
\name{adjust_str}
\alias{adjust_str}
\title{Drop leading and trailing whitespaces, punctuation signs and capitalize strings.}
\usage{
adjust_str(x)
}
\arguments{
\item{x}{A string}
}
\value{
A string
}
\description{
Drop leading and trailing whitespaces, punctuation signs and capitalize strings.
}
\examples{
x <- "cat#!"
}
