% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_multi.R
\name{drop_multi}
\alias{drop_multi}
\title{Drop strings to the left of the first comma}
\usage{
drop_multi(x)
}
\arguments{
\item{x}{A character.}
}
\value{
A character.
}
\description{
Drop strings to the left of the first comma
}
\examples{
x <- "cat, dog"
}
