% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_str.R
\name{match_str}
\alias{match_str}
\title{Match the closest string using the Jaro-Winkler metric from stringdist package}
\usage{
match_str(x, y)
}
\arguments{
\item{x}{A data frame or tibble of strings to match}

\item{y}{A data frame or tibble of strings to use as references}
}
\value{
A tibble!!!!!!!!
}
\description{
Match the closest string using the Jaro-Winkler metric from stringdist package
}
