% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{reader}
\alias{reader}
\title{Import csv or xlsx files}
\usage{
reader(path, head)
}
\arguments{
\item{path}{filepath}

\item{head}{TRUE if the file to import has a first row of headers}
}
\value{
A data frame
}
\description{
Import csv or xlsx files
}
